Zmodyfikowanie kodu tak aby był bardziej czytelny i przyjazny środowisku
